pyproject.toml
[project]
name = "vygil-ai"
version = "1.0.0"
description = "A vigilant guard that watches and monitors"
requires-python = ">=3.9"

dependencies = [
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1", 
    "openai>=1.3.0",
    "anthropic>=0.8.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

